requirements.txt
Django==1.5.1
South==0.8.1
django-extensions==1.1.1
django-smart-selects==1.0.9
psycopg2==2.5.1
six==1.3.0
wsgiref==0.1.2
